requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0.post0
gunicorn==20.1.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.43
Werkzeug==2.2.2
zipp==3.10.0
